fix for ffi enum types differing on different platforms

+++ src/gradient.rs
@@ -238,7 +238,7 @@ impl<T: GradientType> Gradient<T> {
 
     #[inline]
     pub fn has_matrix(&self) -> bool {
-        self.impl_().matrixType as i32 != ffi::BLMatrix2DType::BL_MATRIX2D_TYPE_IDENTITY
+        self.impl_().matrixType as i32 != ffi::BLMatrix2DType::BL_MATRIX2D_TYPE_IDENTITY as i32
     }
 
     #[inline]
